requirements.txt
langchain
langchain_groq
langchain_community
huggingface_hub
python-dotenv
streamlit